Added support to Windows logins
Added some settings in the STS to support the Windows login:
AutomaticWindowsLogin: if true, when authentication is required and if the request comes from the same network as the STS, no login screen is shown and the Windows login automatically takes place;
IncludeWindowsGroups: if user uses Windows authentication, should we load the groups from Windows?
WindowsGroupsPrefix: if we load the groups from Windows, load only the groups that start with this prefix
WindowsGroupsOURoot: if we load the groups from Windows, load only the groups under this OU.

Added user photo to the profile: when the user logins through Windows it is read from the Active Directory, otherwise the "picture" claim of the external login is checked.
When the user photo needs to be shown, we first check if the "picture" claim exists, otherwise we use the Gravatar picture as before.